pyproject.toml
[project]
name = "trans-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.6.0",
    "litellm>=1.74.3",
    "openai-agents>=0.2.0",
    "python-dotenv>=1.1.1",
]
